% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_group_size}
\alias{calc_group_size}
\title{Calculate Mean Group Size from Captures}
\usage{
calc_group_size(cam_captures)
}
\arguments{
\item{cam_captures}{Data frame of processed camera captures.}
}
\value{
Numeric mean group size.
}
\description{
Calculate Mean Group Size from Captures
}
