% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_in_triangle}
\alias{calc_in_triangle}
\title{Check if Point is Inside Triangle (Viewshed)}
\usage{
calc_in_triangle(p_viewshed, point)
}
\arguments{
\item{p_viewshed}{Matrix (3x2) of triangle vertices.}

\item{point}{Vector (length 2) of the point to test.}
}
\value{
Logical (TRUE/FALSE).
}
\description{
Check if Point is Inside Triangle (Viewshed)
}
