% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_intersects}
\alias{calc_intersects}
\title{Calculate Intersection of Trajectory and Camera Viewshed}
\usage{
calc_intersects(p_viewshed, p_animal, speed, t)
}
\arguments{
\item{p_viewshed}{Matrix (3x2) of camera triangle vertices.}

\item{p_animal}{Matrix (Nx2) of animal trajectory coordinates.}

\item{speed}{Vector of animal speeds at each step.}

\item{t}{Vector of time stamps.}
}
\value{
A list containing:
\itemize{
\item Coordinates of intersection points.
\item \code{t_stay}: Duration of each event.
\item \code{in_cam_all}: Boolean vector of presence.
\item \code{encounter}: Encounter IDs.
\item \code{t_in}, \code{t_out}: Entry and exit timestamps.
}
}
\description{
The core geometric engine. It iterates through an animal's movement path to:
\enumerate{
\item Check if segments cross the camera viewshed boundaries.
\item Check if points lie inside the viewshed.
\item Calculate exact entry/exit times and total duration (\code{t_stay}).
}
}
