% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_tri_area}
\alias{calc_tri_area}
\title{Calculate Triangle Area}
\usage{
calc_tri_area(x, y)
}
\arguments{
\item{x}{Vector of three X coordinates.}

\item{y}{Vector of three Y coordinates.}
}
\value{
Numeric area.
}
\description{
Calculate Triangle Area
}
