% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_tri_signs}
\alias{calc_tri_signs}
\title{Calculate Half-Plane Signs}
\usage{
calc_tri_signs(p1, p2, p3)
}
\arguments{
\item{p1, p2, p3}{Points (vectors of length 2).}
}
\value{
Numeric value indicating relative position.
}
\description{
Helper for Barycentric coordinate calculations.
}
