% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Cell_bias_xy}
\alias{Cell_bias_xy}
\title{Modify Landscape Cells by Coordinate (Helper)}
\usage{
Cell_bias_xy(lscape_data, Index_x, Index_y, Dir, Kappa_in, Road_ID)
}
\description{
Manually bias specific landscape cells based on X/Y ranges.
}
