% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_activity_mat}
\alias{create_activity_mat}
\title{Initialize Activity Matrix}
\usage{
create_activity_mat(study_design, animalxy.0)
}
\arguments{
\item{study_design}{A list containing \code{t_steps}.}

\item{animalxy.0}{A data frame of initial animal positions.}
}
\value{
The \code{animalxy.0} data frame with an added \code{activity_mat} column.
}
\description{
Creates a placeholder list for animal activity (active/inactive) states.
Currently defaults to "active" (1) for all time steps.
}
