% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_animal_data}
\alias{get_animal_data}
\title{Calculate True Space Use (Ground Truth)}
\usage{
get_animal_data(animalxy.all, lscape_defs)
}
\arguments{
\item{animalxy.all}{Trajectory data frame.}

\item{lscape_defs}{Landscape definition data frame.}
}
\value{
A summary data frame of time spent per landscape unit.
}
\description{
Aggregates the actual time animals spent in each landscape type based on their
simulated trajectories.
}
