% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covariate_info}
\alias{get_covariate_info}
\title{Extract Unique Covariate Levels}
\usage{
get_covariate_info(data_in, cov_name)
}
\arguments{
\item{data_in}{A data frame.}

\item{cov_name}{Character string of the column name to extract.}
}
\value{
A vector of unique, non-NA labels.
}
\description{
Extract Unique Covariate Levels
}
