% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_ABM}
\alias{plot_ABM}
\title{Plot Simulation Trajectories and Camera Locations}
\usage{
plot_ABM(study_design, cam_design, cam_locs, animalxy.all)
}
\arguments{
\item{study_design}{A list of simulation parameters including \code{bounds} and \code{dx}.}

\item{cam_design}{A list containing camera design details, specifically \code{ncam}.}

\item{cam_locs}{A data frame of camera locations/vertices.}

\item{animalxy.all}{A data frame of animal trajectories (output from \code{ABM_sim}).}
}
\value{
A \code{ggplot} object.
}
\description{
Generates a spatial plot showing the landscape grid, animal trajectories, and
camera trap polygons.
}
