% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_functions.R
\name{fit.model.mcmc.REST.cov}
\alias{fit.model.mcmc.REST.cov}
\title{Fit Landscape Density Model using MCMC}
\usage{
fit.model.mcmc.REST.cov(
  study_design,
  cam_design,
  cam_locs,
  gamma_start,
  kappa_start,
  gamma_prior_var = 10^4,
  kappa_prior_var = 10^4,
  gamma_tune,
  kappa_tune,
  encounter_data_in,
  stay_time_data_in
)
}
\arguments{
\item{study_design}{A list of simulation/study parameters. Must include:
\itemize{
\item \code{n_iter}: Number of MCMC iterations.
\item \code{dx}, \code{dy}: Grid cell dimensions.
\item \code{t_steps}: Number of time steps.
\item \code{covariate_labels}: Names of the covariates (e.g., speed categories).
\item \code{num_covariates}: Integer count of covariates.
\item \code{Z}: Covariate design matrix.
\item \code{q}: Total number of grid cells.
}}

\item{cam_design}{A list containing camera design details, specifically
\code{cam_A} (area of camera view).}

\item{gamma_start}{Vector of initial values for density coefficients.}

\item{kappa_start}{Vector of initial values for staying time coefficients.}

\item{gamma_prior_var}{Numeric, variance for the Normal prior on gamma.}

\item{kappa_prior_var}{Numeric, variance for the Normal prior on kappa.}

\item{gamma_tune}{Vector of initial tuning variances for gamma proposals.}

\item{kappa_tune}{Vector of initial tuning variances for kappa proposals.}

\item{encounter_data_in}{A data frame containing observed encounters}

\item{stay_time_data_in}{A data frame containing observed staying time for each encounter.}

\item{kappa_prior_mu}{Vector of means for the Normal prior on kappa.}
}
\value{
A list containing MCMC chains and diagnostics:
\itemize{
\item \code{accept}: Matrix of acceptance (1/0) events for diagnostics.
\item \code{gamma}: Matrix of posterior samples for density coefficients.
\item \code{kappa}: Matrix of posterior samples for staying time coefficients.
\item \code{tot_u}: Vector of estimated total abundance/density across iterations.
}
}
\description{
Fit Landscape Density Model using MCMC
}
