% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_speeds}
\alias{sample_speeds}
\title{Sample Landscape Indices for Camera Placement}
\usage{
sample_speeds(cam.dist.prop = NULL, lscape_speeds)
}
\arguments{
\item{cam.dist.prop}{A list containing design parameters (equivalent to \code{cam_design}).}

\item{lscape_speeds}{A data frame of landscape definitions.}
}
\value{
A vector of integer indices.
}
\description{
Helper function to select grid cell indices for cameras based on a specific
design strategy (Random, Road Bias, or Speed-stratified).
}
